requirements.txt
fastapi
uvicorn[standard]
ollama
langchain
langchain-core
deepeval
pytest
prometheus_client
langfuse
python-dotenv
pyyaml
